requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
faiss-gpu>=1.7.4  # 可选，如果使用GPU
pypdf>=3.17.0
numpy>=1.24.0
PyYAML>=6.0
torch>=2.0.0
tqdm>=4.66.0